Revert "Factor out some of `h.accounts.schemas` into separate modules"
