--- h/schemas/forms/accounts/__init__.py
@@ -1,16 +1,0 @@
-# -*- coding: utf-8 -*-
-
-from __future__ import unicode_literals
-
-from h.schemas.forms.accounts.edit_profile import EditProfileSchema
-from h.schemas.forms.accounts.forgot_password import ForgotPasswordSchema
-from h.schemas.forms.accounts.login import LoginSchema
-from h.schemas.forms.accounts.reset_password import ResetPasswordSchema
-
-
-__all__ = (
-    'EditProfileSchema',
-    'ForgotPasswordSchema',
-    'LoginSchema',
-    'ResetPasswordSchema',
-)
